+++ README.md
@@ -114,7 +114,7 @@ All dependencies for `curvenote` are included in this repository (a monorepo!).
 **Packages:**
 
 - `ui-providers`: React providers for references and site configuration
-- `myst-util-to-react`: expose MyST content as an article, built with React
+- `myst-to-react`: expose MyST content as an article, built with React
 - `site`: components and utilities for building React and Remix sites
 - `site-common`: utilities and types that are used in the CLI, APIs and sites, does not include React
 - `icons`: icons used throughout our projects, built for React
