+++ apps/web/package.json
@@ -37,7 +37,7 @@
     "date-fns": "^2.28.0",
     "lodash.throttle": "^4.1.1",
     "myst-spec": "^0.0.4",
-    "myst-util-to-react": "*",
+    "myst-to-react": "*",
     "mystjs": "^0.0.13",
     "nanoid": "^3.3.2",
     "node-cache": "^5.1.2",
