+++ packages/myst-to-react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "myst-util-to-react",
+  "name": "myst-to-react",
   "version": "0.1.13",
   "main": "./src/index.tsx",
   "types": "./src/index.tsx",
