+++ packages/myst-to-react/src/components/CopyIcon.tsx
