+++ packages/myst-to-react/src/extensions/chemicalFormula.tsx
