+++ packages/myst-to-react/src/footnotes.tsx
