+++ packages/myst-to-react/src/image.tsx
