+++ packages/myst-to-react/src/math.tsx
