+++ packages/myst-to-react/src/output/error.tsx
