+++ packages/myst-to-react/src/output/safe.tsx
