+++ packages/myst-to-react/src/output/stream.tsx
