+++ packages/site/src/pages/Article.tsx
@@ -1,7 +1,7 @@
 import type { PageLoader } from '@curvenote/site-common';
 import { ReferencesProvider } from '@curvenote/ui-providers';
 import { useLoaderData } from '@remix-run/react';
-import { Bibliography } from 'myst-util-to-react';
+import { Bibliography } from 'myst-to-react';
 import { ContentBlocks, FooterLinksBlock, FrontmatterBlock } from '../components';
 import { ErrorDocumentNotFound } from './ErrorDocumentNotFound';
